requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
Django==4.2
django-crispy-forms==2.0
django-taggit==4.0.0
idna==3.4
Pillow==9.5.0
requests==2.31.0
sqlparse==0.4.4
stripe==5.4.0
urllib3==2.0.2
